simplify api to system and lvm
Simpify api to just system and lvm interfaces and have the rest of the
constructs are structs.

Interfaces: System, VolumeManager
Structs: Disk, Partition, VG, LV, PV

This allows for all the structs and the corresponding methods to be
reused by mock implementation. All the platform specific methods are
modeled in the interfaces and can be implemented per platform.

Signed-off-by: Ravi Chamarthy <ravchama@cisco.com>