Zero start and end of a to-be-created partition addPartitionSet.
When partitions are added to a disk, udev events will scan the
newly created block devices for signatures and act on those signatures.
This means that old data might be acted on.  This often manifests itself
as lvm finding old data but all sorts of other unexpected things can
also occur.

The solution is to make sure that these new block devices have 1MiB
worth of zeros at the beginning and end before releasing the lock.