Connect megaraid to linux System.
This connects the linux disko.System() interface with a storCli
implementation.  The end result is that now we get improved answers
for the DiskType of a Disk.  If the disk is a virtual megaraid drive
whose DiskGroup is entirely made up of ssd, then you will now see
that its DiskTypetype is SSD.

It is hooked up through go-cache to avoid repeated queries of 'storcli'.
That isn't *slow*, but not fast either.  The assumption is that
those things dont' change very often.  The timeout is set to 5 minutes
here.