fix CreatePV to use name and path
CreatePV accepts a path but thinks it can accept a name and uses the
param as a name to later find the PV it just created. The scan fails to
match and CreatePV always returns an error saying it did not find the PV
it just successfully created.

Fix: accept name or path, if name is given make a path and if path is
given make a name and then use both appropriately

Signed-off-by: Ravi Chamarthy <ravchama@cisco.com>