Call partx --add and --delete after adding or removing partitions.
If a disk is mounted, then adding a partition will not result in udev
correctly signalling the kernel to reread the partition table.
This is because udev events just call 'blkrereadpt' (reread the whole
partition table).

We use 'partx' here, but we could just as well use 'addpart' and
'delpart'.  Even better would be to call the ioctl directly.

Fixes #53.