Fix storcli disk parsing when a disk is 'Foreign' (F).
storcli output may contain a disk with a 'F' in its DiskGroup.
This indicates 'Foreign'.  Previously we had stored '-' (not part of a
drive group) as -1.  For now we will store Foreign as -2.

At some point it'd be nice to have constants and such for this,
but for now this should suffice.