Speed up ScanVGs by only scanning PVs and LVs at most once.
Instead of making calls to ScanPVs and ScanLVs when walking the
response from ScanVGs, we only do those once.

The ScanPV call is very slow if there are a lot of block devices
(including lvm block devices).

Overall, this results in considerable speedup.

                 |  vm  | hardware
     with-change | 1.22 | 1.35
     without     | 1.74 | 2.99

They're by no means scientific numbers, but it is a considerable win.