Change addPartitionSet to use addpart instead of partx, release lock first.
Since we know the size of the partition we are adding, we can call
addpart instead of partx.  partx essentially just opens the block
device, finds the size, and then calls BLKPG ioctl.

The real change here is first closing the block device,
releasing the lock, and then calling addpart.