Zero the partition table before writing a protective MBR.
Previously there was a bug where calling newProtectiveMBR would fail
if data in the partition table had invalid values in the bootable flag.

What we do here is just set the bytes of the partition table to a fully
zeroed MBR before then creating the partitions correctly.
There is not a method exposed in MBR library to set the bootable flag,
so we just zero the whole area (NonBootable is represented by 0x00
so this is fine).