Add limited support for MBR/DOS partition tables.
This adds code to write MBR partitions.  Ideally we only write these in
update cases.

It would be nice to have a 'convert' of some sort that would convert MBR
into GPT.  That would have to take into account the fact that we can't
yet read extended partitions.

Also, if a partition table is messed up then addPartitions will fail
because of the check call.  In that event, you'd need to disk.Wipe()
first.  Kind of annoying.  I think the right way to support writing
a new MBR would look a lot like newProtectiveMBR() that reads the
disk and just wipes the mbr section of the first sector, but leaves
the rest intact.

Also adds Table to Disk to store GPT or MBR or None.