Add 'ptimage' tool.
This creates a filesystem on an image in whatever extra space there is.
And copies a directory contents to it.

You can use '--execute' to run a user-provided command:

 $ ./ptimg --execute --copy=my-data/ xx.img ls -l MOUNT_POINT

the environment of the invoked command will have contain:
 MOUNT_POINT - the location of the mounted filesystem.
 PART_DEV - the path to the partition (/dev/nbd0p1)
 BLOCK_DEV - the path to the block device (/dev/nbd0)

Additionally, occurances of the strings above surrounded by '@'
will be replaced in the command executed.
So if the command provided is:
  blkid @PART_DEV@
Then the command executed will be:
  blkid /dev/nbd0p1

To use the environment variable instead, provide the command:
  sh -c 'blkid $PART_DEV'