Add properties to a disk to identify 'Ephemeral'. Identify Azure Ephemeral.
Cloud systems often have 'ephemeral' devices, which are backed by disks
in the local system.  These are generally faster or cheaper but less
reliable than other disks.

This adds a Disks.Properties{} that has a list of properties.
The only property at this point is Ephemeral.

Also identify ephemeral devices on Azure.