Support failure if integration tests cannot run.
Previously intergration tests would just Skip() if they did not have the
necessary tools to run.  That could lead to a situation where we don't
realize the tests are not running and happily trust c-i that was not
running integration tests.

Here, we let the behavior be controlled by DISKO_INTEGRATION environment
variable.
  * allow-skip (default): act as before (run if possible, or skip)
  * skip: always skip.
  * run: raise errors and fail if the conditions are not met.

Also, we set DISKO_INTEGRATION=run in c-i, so failures will fail.