Use --wipesignatures=y on lvcreate.
I had assumed that '--zero=y' was a superset of '--wipesignatures=y'.
That seems to not be the case.  It seems some signatures might live
outside of the area that is zero'd.  So, wipe them.

Add a test called TestRootLVMRecreate that creates some volumes
and then deletes them and then creates them again.  The hope was
to catch the bug where mkfs complained about crypt or fsdata being
present.