Add a Read-Only property to indicate disks to which disko can't wipe
On Linux, block devices which are not writable have the sysfs attribute
'ro' set to '1', otherwise '0'.  When scanning disks, check the 'ro'
sysfs attribute, and if '1' then set the READ-ONLY property on the
Disk object.  This can be used to not consider disks which are marked
read-only for wiping; the writes fail and the command aborts.