Add a test that that creates and verifies an MBR Table.
The interface to creating a MBR is a bit hokey, but it does work.
If a disko.Disk has a Table type of MBR, then CreatePartition
will create a partition table of MBR type.

Definitely this could use some work, as the filled-out ID and Name
are ignored, and the PartitionType gets converted, but it does
work.