Wipe luks2 headers after lvcreate.
Versions of libblkid, before 2.33 would not completely wipe luks 2
headers.  After we create a thick LV, wipe any remaining headers
ourselves.

This should not be that big of a problem, but when it was mixed with
a liberal running of 'cryptsetup luksUUID' after LV creation things
went awry.  cryptsetup will automatically repair a broken header
that it finds, and it would find such a header because lvm did not
completely wipe it.  To be clear, 'cryptsetup luksUUID' is not
a read-only operation.

Also complicating things was the fact that we were creating luks 2
volumes and then running with an old LVM.